Server/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
SQLAlchemy==1.4.46
psycopg2-binary==2.9.5
PyJWT==2.6.0
Werkzeug==2.2.2
influxdb==5.3.1
python-dotenv==0.21.1